pyproject.toml
[project]
name = "trend-following-mcp"
version = "0.1.0"
description = "Trend Following Investment MCP Server with Stock Analysis and Trading Signals"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.0.0",
    "yfinance>=0.2.59",
    "pandas>=2.2.3",
    "numpy>=2.2.5",
    "matplotlib>=3.10.1",
    "seaborn>=0.13.2",
    "ta>=0.11.0",
    "scikit-learn>=1.4.0",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.40",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "pydantic>=2.6.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "schedule>=1.2.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "ruff>=0.11.8",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py313"
line-length = 88
select = ["E", "F", "I", "N", "W", "B", "C4", "UP", "ARG", "SIM", "TCH", "TID", "Q"]
ignore = ["E501", "B008", "C901"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["B011"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.black]
target-version = ['py313']
line-length = 88
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''
